Add files via upload
This is a small web development project that I am working on to enhance my skills in building and connecting databases to web applications. The project utilizes HTML, CSS, and JavaScript for the frontend development. I am learning how to seamlessly integrate a database to store and manage data for the application.

Technologies Used
HTML: For structuring the content of the web pages.
CSS: For styling and layout design to create an appealing user interface.
JavaScript: For adding interactive features and dynamic content.
Database Integration: Learning and implementing database connectivity to handle data storage and retrieval.
Features
Responsive Design: The web pages are designed to be responsive and user-friendly across different devices.
Dynamic Content: Use of JavaScript to create a dynamic user experience.
Database Connectivity: Integration of a database to manage and store user data efficiently.
Goals
To gain hands-on experience in web development using HTML, CSS, and JavaScript.
To understand the fundamentals of database integration with web applications.
To build a functional and interactive web application with data management capabilities.